requirement.txt
requests==2.28.1
APScheduler==3.9.1
arrow==1.2.2
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
certifi==2022.5.18.1
chardet2==2.0.3
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==3.2.13
django-apscheduler==0.6.2
django-cors-headers==3.10.1
django-extensions==3.2.0
django-redis==5.2.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
kafka-python==2.0.2
keyring==18.0.1
numpy==1.23.1
oauthlib==3.1.0
opencv-python==4.6.0.66
packaging==21.3
plucky==0.4.3
py3Fdfs==2.2.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==3.13
pyzmq==23.2.1
redis==4.3.3
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uWSGI==2.0.20
WebOb==1.2b3
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
tcping==0.1.1rc1
django-db-reconnect==0.0.1
channels==3.0.5
channels-redis==4.0.0
sensecam-control==0.1.0
Pillow==9.0.0
pycrypto==2.6.1
paramiko==2.11.0